Move client device info
